# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Cloud IDE API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type FileEntryType = typeof FileEntryType[keyof typeof FileEntryType];


export const FileEntryType = {
  file: 'file',
  directory: 'directory',
} as const;

export interface FileEntry {
  name: string;
  path: string;
  type: FileEntryType;
  /** @nullable */
  size?: number | null;
  /** @nullable */
  children?: FileEntry[] | null;
}

export interface FileContent {
  path: string;
  content: string;
}

export interface FileWrite {
  path: string;
  content: string;
}

export interface FileRename {
  oldPath: string;
  newPath: string;
}

export interface DirectoryCreate {
  path: string;
}

export type CodeRunLanguage = typeof CodeRunLanguage[keyof typeof CodeRunLanguage];


export const CodeRunLanguage = {
  python: 'python',
  node: 'node',
  bash: 'bash',
} as const;

export interface CodeRun {
  language: CodeRunLanguage;
  code: string;
  /** @nullable */
  filename?: string | null;
}

export type AiRequestAction = typeof AiRequestAction[keyof typeof AiRequestAction];


export const AiRequestAction = {
  generate: 'generate',
  fix: 'fix',
  explain: 'explain',
  refactor: 'refactor',
} as const;

export interface AiRequest {
  action: AiRequestAction;
  /** @nullable */
  code?: string | null;
  /** @nullable */
  prompt?: string | null;
  /** @nullable */
  language?: string | null;
  /** @nullable */
  filename?: string | null;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterBody {
  username: string;
  password: string;
  /** @nullable */
  email?: string | null;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface AuthUser {
  id: string;
  username: string;
  email: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: AuthUser;
}

export type DeployConfigType = typeof DeployConfigType[keyof typeof DeployConfigType];


export const DeployConfigType = {
  static: 'static',
  node: 'node',
  python: 'python',
} as const;

export interface DeployConfig {
  type: DeployConfigType;
  buildCommand: string;
  startCommand: string;
  port: number;
  entryPoint: string;
}

export interface Project {
  id: string;
  name: string;
  description: string;
  ownerId: string;
  createdAt: string;
  updatedAt: string;
  deployConfig: DeployConfig;
}

export interface CreateProjectBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  ownerId?: string | null;
}

/**
 * @nullable
 */
export type UpdateProjectBodyDeployConfig = { [key: string]: unknown } | null;

export interface UpdateProjectBody {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  deployConfig?: UpdateProjectBodyDeployConfig;
}

/**
 * @nullable
 */
export type DeployStartBodyType = typeof DeployStartBodyType[keyof typeof DeployStartBodyType] | null;


export const DeployStartBodyType = {
  static: 'static',
  node: 'node',
  python: 'python',
} as const;

export interface DeployStartBody {
  /** @nullable */
  type?: DeployStartBodyType;
  /** @nullable */
  buildCommand?: string | null;
  /** @nullable */
  startCommand?: string | null;
  /** @nullable */
  entryPoint?: string | null;
}

export type DeploymentStatusStatus = typeof DeploymentStatusStatus[keyof typeof DeploymentStatusStatus];


export const DeploymentStatusStatus = {
  idle: 'idle',
  building: 'building',
  running: 'running',
  stopped: 'stopped',
  failed: 'failed',
} as const;

export interface DeploymentStatus {
  projectId: string;
  status: DeploymentStatusStatus;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  pid?: number | null;
  /** @nullable */
  port?: number | null;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  stoppedAt?: string | null;
}

export interface DeploymentLogs {
  projectId: string;
  type: string;
  logs: unknown[];
}

export interface DevOpsAnalyzeBody {
  projectId: string;
  /** @nullable */
  issue?: string | null;
  /** @nullable */
  autoFix?: boolean | null;
}

export interface FileContextEntry {
  path: string;
  content: string;
  language: string;
  size: number;
  imports: string[];
  functions: string[];
  classes: string[];
}

export interface ProjectContext {
  totalFiles: number;
  totalSize: number;
  files: FileContextEntry[];
  structure: string;
  generatedAt: string;
}

export type AgentRequestMode = typeof AgentRequestMode[keyof typeof AgentRequestMode];


export const AgentRequestMode = {
  builder: 'builder',
  debugger: 'debugger',
  reviewer: 'reviewer',
  auto: 'auto',
} as const;

export interface AgentRequest {
  mode: AgentRequestMode;
  task: string;
  /** @nullable */
  targetFile?: string | null;
  /** @nullable */
  errorOutput?: string | null;
  autoRun?: boolean;
}

export interface ProjectMemory {
  projectName: string;
  description: string;
  architecture: string[];
  namingConventions: string[];
  codingStyle: string[];
  dependencies: string[];
  notes: string[];
  updatedAt: string;
}

export type ListFilesParams = {
path?: string;
};

export type ReadFileParams = {
path: string;
};

export type DeleteFileParams = {
path: string;
};

export type GetDeploymentLogsParams = {
type?: GetDeploymentLogsType;
};

export type GetDeploymentLogsType = typeof GetDeploymentLogsType[keyof typeof GetDeploymentLogsType];


export const GetDeploymentLogsType = {
  build: 'build',
  runtime: 'runtime',
  all: 'all',
} as const;