# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Cloud IDE API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /fs/read:
    get:
      operationId: readFile
      tags: [filesystem]
      summary: Read file content
      parameters:
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileContent"
        "404":
          description: File not found
  /fs/write:
    post:
      operationId: writeFile
      tags: [filesystem]
      summary: Write or update file content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FileWrite"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /auth/register:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /deploy/{projectId}/start:
    post:
      operationId: startDeployment
      tags: [deploy]
      summary: Start deployment (SSE)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeployStartBody"
      responses:
        "200":
          description: SSE stream of deploy steps
          content:
            text/event-stream:
              schema:
                type: string
  /deploy/{projectId}/stop:
    post:
      operationId: stopDeployment
      tags: [deploy]
      summary: Stop a running deployment
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /deploy/{projectId}/status:
    get:
      operationId: getDeploymentStatus
      tags: [deploy]
      summary: Get deployment status
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deployment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentStatus"
  /deploy/{projectId}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deploy]
      summary: Get deployment logs
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [build, runtime, all]
      responses:
        "200":
          description: Deployment logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentLogs"
...
components:
  schemas:
...
    FileContent:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
    FileWrite:
      type: object
      properties:
        path:
          type: string
        content:
          type: string
      required:
        - path
        - content
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
    RegisterBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: ["string", "null"]
      required:
        - username
        - password
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - createdAt
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required:
        - token
        - user
...
    DeployStartBody:
      type: object
      properties:
        type:
          type: ["string", "null"]
          enum: [static, node, python, null]
        buildCommand:
          type: ["string", "null"]
        startCommand:
          type: ["string", "null"]
        entryPoint:
          type: ["string", "null"]
    DeploymentStatus:
      type: object
      properties:
        projectId:
          type: string
        status:
          type: string
          enum: [idle, building, running, stopped, failed]
        url:
          type: ["string", "null"]
        pid:
          type: ["integer", "null"]
        port:
          type: ["integer", "null"]
        startedAt:
          type: ["string", "null"]
        stoppedAt:
          type: ["string", "null"]
      required:
        - projectId
        - status
    DeploymentLogs:
      type: object
      properties:
        projectId:
          type: string
        type:
          type: string
        logs:
          type: array
          items: {}
      required:
        - projectId
        - type
        - logs
...